backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# ASR & ML
mlx-whisper>=0.4.0
pyannote.audio>=3.3.0
torch>=2.1.0
torchaudio>=2.1.0
whisperx>=3.1.0

# LLM
litellm==1.17.9

# Utils
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
watchdog==3.0.0
aiofiles==23.2.1

# Testing
pytest==8.0.0
httpx==0.26.0
